Authorization with Policies

+++ app/Http/Controllers/ThreadsController.php
@@ -111,7 +111,9 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
-      // $thread->replies()->delete();
+
+      $this->authorize('update', $thread);
+
       $thread->delete();
 
       if(request()->wantsJson())
